+++ source/TextArea.h
@@ -323,7 +323,7 @@ private:
 	void extendRangeForStyleMods(int *start, int *end);
 	void redrawLineNumbers(QPainter *painter, bool clearAll);
 	void redrawLineNumbersEx(bool clearAll);
-	void redisplayLine(QPainter *painter, int visLineNum, int leftClip, int rightClip, int leftCharIndex, int rightCharIndex);
+    void redisplayLine(QPainter *painter, int visLineNum, int leftClip, int rightClip, int leftCharIndex, long rightCharIndex);
 	void redisplayLineEx(int visLineNum, int leftClip, int rightClip, int leftCharIndex, int rightCharIndex);
 	int styleOfPos(int lineStartPos, int lineLen, int lineIndex, int dispIndex, int thisChar);
     void drawString(QPainter *painter, int style, int x, int y, int toX, char *string, int nChars);
